pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools-scm>=6.2", "cmake>=3.15", "pybind11>=2.10.0", "numpy"]
build-backend = "setuptools.build_meta"

[project]
name = "wdr-compression-pipeline"
version = "0.1.0"
description = "Hybrid Python/C++ WDR Image Compression Pipeline"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.20.0",
    "pywavelets>=1.3.0",
    "Pillow>=8.0.0",
    "tifffile>=2024.2.12",
    "zarr>=2.16.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
include = ["wdr*"]

